+++ src/cmd/sql_cmd_test.cc
@@ -29,6 +29,7 @@
 #include "sdk/mini_cluster.h"
 #include "sdk/sql_router.h"
 #include "vm/catalog.h"
+#include "test/util.h"
 
 DECLARE_bool(interactive);
 DEFINE_string(cmd, "", "Set cmd");
@@ -41,65 +42,10 @@ namespace openmldb {
 namespace cmd {
 
 ::openmldb::sdk::MiniCluster* mc_;
-inline std::string GenRand() {
-    return std::to_string(rand() % 10000000 + 1);  // NOLINT
-}
 
-inline void ProcessSQLs(sdk::SQLClusterRouter* sr, std::initializer_list<absl::string_view> sqls) {
-    hybridse::sdk::Status status;
-    for (auto sql : sqls) {
-        sr->ExecuteSQL(std::string(sql), &status);
-        EXPECT_TRUE(status.IsOK()) << "running sql=" << sql << " failed: " << status.msg;
-    }
-}
-
-// expect object for ResultSet cell
-struct CellExpectInfo {
-    CellExpectInfo() {}
-
-    CellExpectInfo(const char* buf) : expect_(buf) {}  // NOLINT
-
-    CellExpectInfo(const std::string& expect) : expect_(expect) {}  // NOLINT
-
-    CellExpectInfo(const std::optional<std::string>& expect) : expect_(expect) {}  // NOLINT
-
-    CellExpectInfo(const std::optional<std::string>& expect, const std::optional<std::string>& expect_not)
-        : expect_(expect), expect_not_(expect_not) {}
-
-    // result string for target cell to match exactly
-    // if empty, do not check
-    std::optional<std::string const> expect_;
-
-    // what string target cell string should not match
-    // if empty, do not check
-    std::optional<std::string const> expect_not_;
-};
-
-// expect the output of a ResultSet, first row is schema, all compared in string
-// if expect[i][j].(expect_|expect_not_) is not set, assert will skip
-inline void ExpectResultSetStrEq(const std::vector<std::vector<CellExpectInfo>>& expect, hybridse::sdk::ResultSet* rs) {
-    ASSERT_EQ(expect.size(), rs->Size() + 1);
-    size_t idx = 0;
-    // schema check
-    ASSERT_EQ(expect.front().size(), rs->GetSchema()->GetColumnCnt());
-    for (size_t i = 0; i < expect[idx].size(); ++i) {
-        ASSERT_TRUE(expect[idx][i].expect_.has_value());
-        EXPECT_EQ(expect[idx][i].expect_.value(), rs->GetSchema()->GetColumnName(i));
-    }
-
-    while (++idx < expect.size() && rs->Next()) {
-        for (size_t i = 0; i < expect[idx].size(); ++i) {
-            std::string val;
-            EXPECT_TRUE(rs->GetAsString(i, val));
-            if (expect[idx][i].expect_.has_value()) {
-                EXPECT_EQ(expect[idx][i].expect_.value(), val) << "expect[" << idx << "][" << i << "]";
-            }
-            if (expect[idx][i].expect_not_.has_value()) {
-                EXPECT_NE(expect[idx][i].expect_not_.value(), val) << "expect_not[" << idx << "][" << i << "]";
-            }
-        }
-    }
-}
+using test::GenRand;
+using test::ProcessSQLs;
+using test::ExpectResultSetStrEq;
 
 struct CLI {
     ::openmldb::sdk::DBSDK* cs = nullptr;
