+++ src/nameserver/system_table_test.cc
@@ -14,18 +14,22 @@
  * limitations under the License.
  */
 
+#include "nameserver/system_table.h"
+
 #include <gflags/gflags.h>
 #include <sched.h>
 #include <unistd.h>
-#include <vector>
-#include <string>
 
+#include <string>
+#include <vector>
+
+#include "absl/time/clock.h"
+#include "absl/time/time.h"
 #include "base/glog_wapper.h"
 #include "brpc/server.h"
 #include "client/ns_client.h"
 #include "common/timer.h"
 #include "gtest/gtest.h"
-#include "nameserver/system_table.h"
 #include "rpc/rpc_client.h"
 #include "test/util.h"
 
@@ -58,6 +62,10 @@ TEST_F(SystemTableTest, SystemTable) {
     ASSERT_TRUE(::openmldb::test::StartNS("127.0.0.1:6530", &ns));
     ::openmldb::client::NsClient ns_client("127.0.0.1:6530", "");
     ns_client.Init();
+
+    // wait a while for that ns become leader
+    absl::SleepFor(absl::Seconds(5));
+
     std::vector<::openmldb::nameserver::TableInfo> tables;
     std::string msg;
     ASSERT_TRUE(ns_client.ShowTable("", INTERNAL_DB, false, tables, msg));
@@ -77,7 +85,7 @@ TEST_F(SystemTableTest, SystemTable) {
     tables.clear();
 
     ASSERT_TRUE(ns_client.ShowTable("", INFORMATION_SCHEMA_DB, false, tables, msg));
-    ASSERT_EQ(1, tables.size());
+    ASSERT_EQ(2, tables.size());
     tables.clear();
     ASSERT_FALSE(ns_client.DropDatabase(INFORMATION_SCHEMA_DB, msg));
